separated local zone ui-test from favorite zone tests
